Migrate MainWindow into window.py
