+++ android/src/toga_android/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, MainWindow
+from .app import App
 from .command import Command
 from .fonts import Font
 from .hardware.camera import Camera
@@ -31,7 +31,7 @@ from .widgets.table import Table
 from .widgets.textinput import TextInput
 from .widgets.timeinput import TimeInput
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -41,7 +41,6 @@ def not_implemented(feature):
 __all__ = [
     "App",
     "Command",
-    "MainWindow",
     "not_implemented",
     # Resources
     "dialogs",
@@ -78,6 +77,8 @@ __all__ = [
     "TimeInput",
     # "Tree",
     "WebView",
+    # Windows
+    "MainWindow",
     "Window",
 ]
