+++ cocoa/src/toga_cocoa/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, DocumentApp, MainWindow
+from .app import App, DocumentApp
 from .command import Command
 from .documents import Document
 from .fonts import Font
@@ -35,7 +35,7 @@ from .widgets.table import Table
 from .widgets.textinput import TextInput
 from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -46,7 +46,6 @@ __all__ = [
     "not_implemented",
     "App",
     "DocumentApp",
-    "MainWindow",
     "Command",
     "Document",
     # Resources
@@ -82,6 +81,8 @@ __all__ = [
     "TextInput",
     "Tree",
     "WebView",
+    # Windows,
+    "MainWindow",
     "Window",
 ]
