+++ core/src/toga/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import warnings
 from pathlib import Path
 
-from .app import App, DocumentApp, DocumentMainWindow, MainWindow
+from .app import App, DocumentApp
 
 # Resources
 from .colors import hsl, hsla, rgb, rgba
@@ -44,7 +44,7 @@ from .widgets.textinput import TextInput
 from .widgets.timeinput import TimeInput, TimePicker
 from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import Window
+from .window import DocumentMainWindow, MainWindow, Window
 
 
 class NotImplementedWarning(RuntimeWarning):
@@ -62,8 +62,6 @@ __all__ = [
     # Applications
     "App",
     "DocumentApp",
-    "MainWindow",
-    "DocumentMainWindow",
     # Commands
     "Command",
     "Group",
@@ -112,6 +110,9 @@ __all__ = [
     "Tree",
     "WebView",
     "Widget",
+    # Windows
+    "DocumentMainWindow",
+    "MainWindow",
     "Window",
     # Deprecated widget names
     "DatePicker",
