+++ core/src/toga/app.py
@@ -10,11 +10,9 @@ from collections.abc import Iterator
 from email.message import Message
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, MutableSet, Protocol
-from warnings import warn
 from weakref import WeakValueDictionary
 
 from toga.command import CommandSet
-from toga.documents import Document
 from toga.handlers import wrapped_handler
 from toga.hardware.camera import Camera
 from toga.hardware.location import Location
@@ -22,13 +20,12 @@ from toga.icons import Icon
 from toga.paths import Paths
 from toga.platform import get_platform_factory
 from toga.screens import Screen
-from toga.types import Position, Size
 from toga.widgets.base import Widget
-from toga.window import OnCloseHandler, Window
+from toga.window import MainWindow, Window
 
 if TYPE_CHECKING:
+    from toga.documents import Document
     from toga.icons import IconContentT
-    from toga.types import PositionT, SizeT
 
 # Make sure deprecation warnings are shown by default
 warnings.filterwarnings("default", category=DeprecationWarning)
@@ -104,7 +101,7 @@ class WindowSet(MutableSet[Window]):
 
     def __iadd__(self, window: Window) -> WindowSet:
         # The standard set type does not have a += operator.
-        warn(
+        warnings.warn(
             "Windows are automatically associated with the app; += is not required",
             DeprecationWarning,
             stacklevel=2,
@@ -114,7 +111,7 @@ class WindowSet(MutableSet[Window]):
     def __isub__(self, other: Window) -> WindowSet:
         # The standard set type does have a -= operator, but it takes sets rather than
         # individual items.
-        warn(
+        warnings.warn(
             "Windows are automatically removed from the app; -= is not required",
             DeprecationWarning,
             stacklevel=2,
@@ -184,120 +181,6 @@ class WidgetRegistry:
 
     def _remove(self, id: str) -> None:
         del self._registry[id]
-
-
-class MainWindow(Window):
-    _WINDOW_CLASS = "MainWindow"
-
-    def __init__(
-        self,
-        id: str | None = None,
-        title: str | None = None,
-        position: PositionT | None = None,
-        size: SizeT = Size(640, 480),
-        resizable: bool = True,
-        minimizable: bool = True,
-        content: Widget | None = None,
-        resizeable: None = None,  # DEPRECATED
-        closeable: None = None,  # DEPRECATED
-    ):
-        """Create a new main window.
-
-        :param id: A unique identifier for the window. If not provided, one will be
-            automatically generated.
-        :param title: Title for the window. Defaults to the formal name of the app.
-        :param position: Position of the window, as a :any:`toga.Position` or tuple of
-            ``(x, y)`` coordinates, in :ref:`CSS pixels <css-units>`.
-        :param size: Size of the window, as a :any:`toga.Size` or tuple of ``(width,
-            height)``, in :ref:`CSS pixels <css-units>`.
-        :param resizable: Can the window be resized by the user?
-        :param minimizable: Can the window be minimized by the user?
-        :param content: The initial content for the window.
-        :param resizeable: **DEPRECATED** - Use ``resizable``.
-        :param closeable: **DEPRECATED** - Use ``closable``.
-        """
-        super().__init__(
-            id=id,
-            title=title,
-            position=position,
-            size=size,
-            resizable=resizable,
-            closable=True,
-            minimizable=minimizable,
-            content=content,
-            # Deprecated arguments
-            resizeable=resizeable,
-            closeable=closeable,
-        )
-
-    @property
-    def _default_title(self) -> str:
-        return App.app.formal_name
-
-    @property
-    def on_close(self) -> None:
-        """The handler to invoke before the window is closed in response to a user
-        action.
-
-        Always returns ``None``. Main windows should use :meth:`toga.App.on_exit`,
-        rather than ``on_close``.
-
-        :raises ValueError: if an attempt is made to set the ``on_close`` handler.
-        """
-        return None
-
-    @on_close.setter
-    def on_close(self, handler: OnCloseHandler | None) -> None:
-        if handler:
-            raise ValueError(
-                "Cannot set on_close handler for the main window. "
-                "Use the app on_exit handler instead."
-            )
-
-
-class DocumentMainWindow(Window):
-    def __init__(
-        self,
-        doc: Document,
-        id: str | None = None,
-        title: str | None = None,
-        position: PositionT = Position(100, 100),
-        size: SizeT = Size(640, 480),
-        resizable: bool = True,
-        minimizable: bool = True,
-    ):
-        """Create a new document Main Window.
-
-        This installs a default on_close handler that honors platform-specific document
-        closing behavior. If you want to control whether a document is allowed to close
-        (e.g., due to having unsaved change), override
-        :meth:`toga.Document.can_close()`, rather than implementing an on_close handler.
-
-        :param doc: The document being managed by this window
-        :param id: The ID of the window.
-        :param title: Title for the window. Defaults to the formal name of the app.
-        :param position: Position of the window, as a :any:`toga.Position` or tuple of
-            ``(x, y)`` coordinates.
-        :param size: Size of the window, as a :any:`toga.Size` or tuple of
-            ``(width, height)``, in pixels.
-        :param resizable: Can the window be manually resized by the user?
-        :param minimizable: Can the window be minimized by the user?
-        """
-        self.doc = doc
-        super().__init__(
-            id=id,
-            title=title,
-            position=position,
-            size=size,
-            resizable=resizable,
-            closable=True,
-            minimizable=minimizable,
-            on_close=doc.handle_close,
-        )
-
-    @property
-    def _default_title(self) -> str:
-        return self.doc.path.name
 
 
 def overridable(method):
@@ -381,7 +264,7 @@ class App:
         # 2023-10: Backwards compatibility
         ######################################################################
         if id is not None:
-            warn(
+            warnings.warn(
                 "App.id is deprecated and will be ignored. Use app_id instead",
                 DeprecationWarning,
                 stacklevel=2,
@@ -569,7 +452,7 @@ class App:
     @property
     def id(self) -> str:
         """**DEPRECATED** – Use :any:`app_id`."""
-        warn(
+        warnings.warn(
             "App.id is deprecated. Use app_id instead", DeprecationWarning, stacklevel=2
         )
         return self._app_id
@@ -868,7 +751,7 @@ class App:
     @property
     def name(self) -> str:
         """**DEPRECATED** – Use :any:`formal_name`."""
-        warn(
+        warnings.warn(
             "App.name is deprecated. Use formal_name instead",
             DeprecationWarning,
             stacklevel=2,
