+++ core/tests/window/test_mainwindow.py
