+++ dummy/src/toga_dummy/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, DocumentApp, MainWindow
+from .app import App, DocumentApp
 from .command import Command
 from .documents import Document
 from .fonts import Font
@@ -36,7 +36,7 @@ from .widgets.textinput import TextInput
 from .widgets.timeinput import TimeInput
 from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -47,7 +47,6 @@ __all__ = [
     "not_implemented",
     "App",
     "DocumentApp",
-    "MainWindow",
     "Command",
     "Document",
     "Font",
@@ -84,6 +83,8 @@ __all__ = [
     "TimeInput",
     "Tree",
     "WebView",
+    # Windows
+    "MainWindow",
     "Window",
     # Widget is also required for testing purposes
     # Real backends shouldn't expose Widget.
