+++ dummy/src/toga_dummy/window.py
@@ -112,3 +112,7 @@ class Window(LoggedObject):
     def get_current_screen(self):
         # `window.screen` will return `Secondary Screen`
         return ScreenImpl(native=("Secondary Screen", (-1366, -768), (1366, 768)))
+
+
+class MainWindow(Window):
+    pass
