+++ gtk/src/toga_gtk/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, DocumentApp, MainWindow
+from .app import App, DocumentApp
 from .command import Command
 from .documents import Document
 from .fonts import Font
@@ -31,7 +31,7 @@ from .widgets.table import Table
 from .widgets.textinput import TextInput
 from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -42,7 +42,6 @@ __all__ = [
     "not_implemented",
     "App",
     "DocumentApp",
-    "MainWindow",
     "Command",
     "Document",
     # Resources
@@ -75,6 +74,8 @@ __all__ = [
     "TextInput",
     "Tree",
     "WebView",
+    # Windows
+    "MainWindow",
     "Window",
 ]
