+++ iOS/src/toga_iOS/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, MainWindow
+from .app import App
 from .colors import native_color
 from .command import Command
 from .fonts import Font
@@ -36,7 +36,7 @@ from .widgets.textinput import TextInput
 
 # from .widgets.tree import Tree
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -46,7 +46,6 @@ def not_implemented(feature):
 __all__ = [
     "not_implemented",
     "App",
-    "MainWindow",
     "Command",
     # Resources
     "native_color",  # colors
@@ -80,6 +79,8 @@ __all__ = [
     "TextInput",
     # 'Tree',
     "WebView",
+    # Windows
+    "MainWindow",
     "Window",
 ]
