+++ textual/src/toga_textual/app.py
@@ -3,12 +3,6 @@ import asyncio
 from textual.app import App as TextualApp
 
 from .screens import Screen as ScreenImpl
-from .window import Window
-
-
-class MainWindow(Window):
-    def textual_close(self):
-        self.interface.app.on_exit()
 
 
 class TogaApp(TextualApp):
