+++ textual/src/toga_textual/window.py
@@ -212,3 +212,8 @@ class Window:
 
     def get_image_data(self):
         self.interface.factory.not_implemented("Window.get_image_data()")
+
+
+class MainWindow(Window):
+    def textual_close(self):
+        self.interface.app.on_exit()
