+++ winforms/src/toga_winforms/factory.py
@@ -1,7 +1,7 @@
 from toga import NotImplementedWarning
 
 from . import dialogs
-from .app import App, MainWindow
+from .app import App
 from .command import Command
 from .fonts import Font
 from .icons import Icon
@@ -30,7 +30,7 @@ from .widgets.table import Table
 from .widgets.textinput import TextInput
 from .widgets.timeinput import TimeInput
 from .widgets.webview import WebView
-from .window import Window
+from .window import MainWindow, Window
 
 
 def not_implemented(feature):
@@ -40,7 +40,6 @@ def not_implemented(feature):
 __all__ = [
     "not_implemented",
     "App",
-    "MainWindow",
     "Command",
     # Resources
     "Font",
@@ -72,7 +71,9 @@ __all__ = [
     "TextInput",
     "TimeInput",
     "WebView",
+    # Windows
     "Window",
+    "MainWindow",
 ]
