+++ winforms/src/toga_winforms/window.py
@@ -251,3 +251,15 @@ class Window(Container, Scalable):
         stream = MemoryStream()
         bitmap.Save(stream, ImageFormat.Png)
         return bytes(stream.ToArray())
+
+
+class MainWindow(Window):
+    def winforms_FormClosing(self, sender, event):
+        # Differentiate between the handling that occurs when the user
+        # requests the app to exit, and the actual application exiting.
+        if not self.interface.app._impl._is_exiting:  # pragma: no branch
+            # If there's an event handler, process it. The decision to
+            # actually exit the app will be processed in the on_exit handler.
+            # If there's no exit handler, assume the close/exit can proceed.
+            self.interface.app.on_exit()
+            event.Cancel = True
